requirements.txt
aws-cdk-lib==2.83.1
constructs>=10.0.0,<11.0.0

boto3~=1.26.154
botocore~=1.29.154